rar_common.c: Fix warning for !HAVE_MMAP case
detected in a legacy linux-x86-64-native build after fixing the
warnings that had been introduced in commit 136a3e8

+++ src/rar_common.c
@@ -118,8 +118,8 @@ typedef struct {
 		unsigned int w;
 		unsigned char c[4];
 	} crc;
-	unsigned long long pack_size;
-	unsigned long long unp_size;
+	uint64_t pack_size;
+	uint64_t unp_size;
 	int method;
 	unsigned char blob_hash[20]; // holds an sha1, but could be 'any' hash.
 	// raw_data should be word aligned, and 'ok'
@@ -181,8 +181,8 @@ static void *get_salt(char *ciphertext)
 	} else {
 		char *p = strtokm(NULL, "*");
 		char crc_c[4];
-		unsigned long long pack_size;
-		unsigned long long unp_size;
+		uint64_t pack_size;
+		uint64_t unp_size;
 
 		for (i = 0; i < 4; i++)
 			crc_c[i] = atoi16[ARCH_INDEX(p[i * 2])] * 16 + atoi16[ARCH_INDEX(p[i * 2 + 1])];
@@ -548,7 +548,7 @@ inline static void check_rar(int count)
 				CRC32_t crc;
 				unsigned char crc_out[4];
 				unsigned char plain[0x8000];
-				unsigned long long size = cur_file->unp_size;
+				uint64_t size = cur_file->unp_size;
 				unsigned char *cipher = cur_file->blob;
 
 				/* Use full decryption with CRC check.
